// src/app/Models/calendario/detalleCalendario-model.ts
export class detalleCalendario{
    IdDetalleCalendario: number;
    IdCalendario: number;
    Folio: number;
    Documento: string;
    Descripcion: string;
    Start: Date;
    Endd: Date;
    Title: string;
    Color: string;
    AllDay: number;
    ResizableBeforeStart: number;
    ResizableBeforeEnd: number;
    Draggable: number;
}